feat(condo): DOMA-7247 create country specific parsers
