+++ apps/condo/domains/billing/utils/receiptQRCodeUtils.js
@@ -1,5 +1,4 @@
 const dayjs = require('dayjs')
-const iconv = require('iconv-lite')
 const { get, isNil, map, set } = require('lodash')
 
 const { createInstance } = require('@open-condo/clients/address-service-client')
@@ -17,41 +16,6 @@ const REQUIRED_QR_CODE_FIELDS = ['BIC', 'PayerAddress', 'Sum', 'PersAcc', 'Payee
  * @type {number}
  */
 const DEFAULT_PERIODS_EDGE_DATE = 20
-
-/**
- * @typedef {Object} TQRCodeFields
- * @property {string} BIC
- * @property {string} PayerAddress
- * @property {string} PaymPeriod
- * @property {string} Sum
- * @property {string} PersAcc The resident's account
- * @property {string} PayeeINN
- * @property {string} PersonalAcc The bank account of the receiver
- */
-
-/**
- * @param {string} qrStr The QR code string got from the picture
- * @return {TQRCodeFields}
- */
-function parseRUReceiptQRCode (qrStr) {
-    const matches = /^ST(?<version>\d{4})(?<encodingTag>\d)\|(?<requisitesStr>.*)$/g.exec(qrStr)
-
-    if (!matches) {
-        throw new Error('Invalid QR code')
-    }
-
-    const requisitesStr = get(matches, ['groups', 'requisitesStr'], '')
-
-    // https://encoding.spec.whatwg.org/#koi8-r
-    // https://encoding.spec.whatwg.org/#windows-1251
-    const encodingTag = get(matches, ['groups', 'encodingTag'])
-    const encoding = get({ 1: 'cp1251', 2: 'utf-8', 3: 'koi8-r' }, encodingTag, 'utf-8')
-
-    // Skip decoding for utf string
-    const decodedRequisitesStr = encodingTag === '2' ? requisitesStr : iconv.decode(requisitesStr, encoding)
-
-    return Object.fromEntries(decodedRequisitesStr.split('|').map((part) => part.split('=', 2)))
-}
 
 /**
  * @param {TQRCodeFields} qrCode
@@ -267,7 +231,6 @@ async function findAuxiliaryData (qrCodeFields, errors) {
 
 module.exports = {
     DEFAULT_PERIODS_EDGE_DATE,
-    parseRUReceiptQRCode,
     getQRCodeField,
     getQRCodeFields,
     getQRCodeMissedFields,
